requirements.txt
streamlit
langchain-google-genai
langchain-chroma
langchain-community
python-dotenv
pypdf
chromadb
langchain
langchain-text-splitters